Optimise size of model structs (#2464)
This commit reduces the sizes of many integer fields, or converts data to enums when appropriate:

- Reduce the size of rate_limit fields
- Reduce the size of channel position
- Reduce size of role position
- Use an enum for `auto_archive_duration`
- `DefaultReaction` -> `ForumEmoji`
- Swap `AfkTimeout` for enum
- Reduce `Sticker::sort_value` to u16
- Reduce size of `min`/`max_values`
- Reduce size of `max_uses` and `max_age`
- Reduce size of `ActivityParty::size`
- Reduce size of `Ready::version`
- Reduce `InputText::min/max_length`
- Reduce `size`/`width`/`height` -- Discord filesize limits mean we won't hit the size limit, and therefore won't hit height/width limits, plus discord rejects 0 sized files so we can use `NonZero*`
- Reduce size of `mention_total_limit`
- Reduce size of `RoleSubscriptionData::total_months_subscribed` -- This gives us ~5000 years of role subscription
- Fix Change sizes

+++ src/builder/create_components.rs
@@ -168,9 +168,9 @@ pub struct CreateSelectMenu {
     #[serde(skip_serializing_if = "Option::is_none")]
     placeholder: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    min_values: Option<u64>,
+    min_values: Option<u8>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    max_values: Option<u64>,
+    max_values: Option<u8>,
     #[serde(skip_serializing_if = "Option::is_none")]
     disabled: Option<bool>,
 
@@ -206,13 +206,13 @@ impl CreateSelectMenu {
     }
 
     /// Sets the minimum values for the user to select.
-    pub fn min_values(mut self, min: u64) -> Self {
+    pub fn min_values(mut self, min: u8) -> Self {
         self.min_values = Some(min);
         self
     }
 
     /// Sets the maximum values for the user to select.
-    pub fn max_values(mut self, max: u64) -> Self {
+    pub fn max_values(mut self, max: u8) -> Self {
         self.max_values = Some(max);
         self
     }
@@ -340,13 +340,13 @@ impl CreateInputText {
     }
 
     /// Sets the minimum length required for the input text
-    pub fn min_length(mut self, min: u64) -> Self {
+    pub fn min_length(mut self, min: u16) -> Self {
         self.0.min_length = Some(min);
         self
     }
 
     /// Sets the maximum length required for the input text
-    pub fn max_length(mut self, max: u64) -> Self {
+    pub fn max_length(mut self, max: u16) -> Self {
         self.0.max_length = Some(max);
         self
     }
